restructure web-extension component
