+++ client/src/components/web-extension/Common/Account/index.scss
