+++ client/src/components/web-extension/Common/Grid/index.scss
