+++ client/src/components/web-extension/Common/Header/index.scss
