+++ client/src/components/web-extension/Common/Layout/BottomBar.jsx
