+++ client/src/components/web-extension/Common/SendReceive/index.scss
