+++ client/src/components/web-extension/Dashboard/index.scss
